Moved uvparameter definitions to separate file
