+++ uvdata/uv.py
@@ -7,147 +7,12 @@ import warnings
 import aipy as a
 import ephem
 from astropy.utils import iers
+import uvdata.parameter as uvp
 import uvdata
 
 data_path = op.join(uvdata.__path__[0], 'data')
 
 iers_a = iers.IERS_A.open(op.join(data_path, 'finals.all'))
-
-
-class UVParameter(object):
-    def __init__(self, name, required=True, value=None, spoof_val=None,
-                 form=(), description='', expected_type=np.int, sane_vals=None,
-                 tols=(1e-05, 1e-08)):
-        self.name = name
-        self.required = required
-        # cannot set a spoof_val for required parameters
-        if not self.required:
-            self.spoof_val = spoof_val
-        self.value = value
-        self.description = description
-        self.form = form
-        if self.form == 'str':
-            self.expected_type = str
-        else:
-            self.expected_type = expected_type
-        self.sane_vals = sane_vals
-        if np.size(tols) == 1:
-            # Only one tolerance given, assume absolute, set relative to zero
-            self.tols = (0, tols)
-        else:
-            self.tols = tols  # relative and absolute tolerances to be used in np.isclose
-
-    def __eq__(self, other):
-        if isinstance(other, self.__class__):
-            # only check that value is identical
-            isequal = True
-            if not isinstance(self.value, other.value.__class__):
-                isequal = False
-            if isinstance(self.value, np.ndarray):
-                if self.value.shape != other.value.shape:
-                    isequal = False
-                elif not np.allclose(self.value, other.value,
-                                     rtol=self.tols[0], atol=self.tols[1]):
-                    isequal = False
-            else:
-                str_type = False
-                if isinstance(self.value, (str, unicode)):
-                    str_type = True
-                if isinstance(self.value, list):
-                    if isinstance(self.value[0], str):
-                        str_type = True
-
-                if not str_type:
-                    try:
-                        if not np.isclose(np.array(self.value),
-                                          np.array(other.value),
-                                          rtol=self.tols[0], atol=self.tols[1]):
-                            isequal = False
-                    except:
-                        print(self.value, other.value)
-                        isequal = False
-                else:
-                    if self.value != other.value:
-                        if not isinstance(self.value, list):
-                            if self.value.replace('\n', '') != other.value.replace('\n', ''):
-                                isequal = False
-                        else:
-                            isequal = False
-
-            return isequal
-        else:
-            return False
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
-
-    def apply_spoof(self, *args):
-        self.value = self.spoof_val
-
-    def expected_size(self, dataobj):
-        # Takes the form of the parameter and returns the size
-        # expected, given values in the UVData object
-        if self.form == 'str':
-            return self.form
-        elif isinstance(self.form, np.int):
-            # Fixed size, just return the form
-            return self.form
-        else:
-            # Given by other attributes, look up values
-            esize = ()
-            for p in self.form:
-                if isinstance(p, np.int):
-                    esize = esize + (p,)
-                else:
-                    val = getattr(dataobj, p)
-                    if val is None:
-                        raise ValueError('Missing UVData parameter {p} needed to '
-                                         'calculate expected size of parameter'.format(p=p))
-                    esize = esize + (val,)
-            return esize
-
-    def sanity_check(self):
-        # A quick method for checking that values are sane
-        # This needs development
-        sane = False  # Default to insanity
-        if self.sane_vals is None:
-            sane = True
-        else:
-            testval = np.mean(np.abs(self.value))
-            if (testval >= self.sane_vals[0]) and (testval <= self.sane_vals[1]):
-                sane = True
-        return sane
-
-
-class AntPositionParameter(UVParameter):
-    def apply_spoof(self, uvdata):
-        self.value = np.zeros((len(uvdata.antenna_indices), 3))
-
-
-class ExtraKeywordParameter(UVParameter):
-    def __init__(self, name, required=False, value={}, spoof_val={},
-                 description=''):
-        self.name = name
-        self.required = required
-        # cannot set a spoof_val for required parameters
-        if not self.required:
-            self.spoof_val = spoof_val
-        self.value = value
-        self.description = description
-
-
-class AngleParameter(UVParameter):
-    def degrees(self):
-        if self.value is None:
-            return None
-        else:
-            return self.value * 180. / np.pi
-
-    def set_degrees(self, degree_val):
-        if degree_val is None:
-            self.value = None
-        else:
-            self.value = degree_val * np.pi / 180.
 
 
 class UVData(object):
@@ -156,117 +21,118 @@ class UVData(object):
 
     def __init__(self):
         # add the UVParameters to the class
-        self._Ntimes = UVParameter('Ntimes', description='Number of times')
-        self._Nbls = UVParameter('Nbls', description='number of baselines')
-        self._Nblts = UVParameter('Nblts', description='Ntimes * Nbls')
-        self._Nfreqs = UVParameter('Nfreqs', description='number of frequency channels')
-        self._Npols = UVParameter('Npols', description='number of polarizations')
+        self._Ntimes = uvp.UVParameter('Ntimes', description='Number of times')
+        self._Nbls = uvp.UVParameter('Nbls', description='number of baselines')
+        self._Nblts = uvp.UVParameter('Nblts', description='Ntimes * Nbls')
+        self._Nfreqs = uvp.UVParameter('Nfreqs', description='number of frequency channels')
+        self._Npols = uvp.UVParameter('Npols', description='number of polarizations')
 
         desc = ('array of the visibility data, size: (Nblts, Nspws, Nfreqs, '
                 'Npols), type = complex float, in units of self.vis_units')
-        self._data_array = UVParameter('data_array', description=desc,
-                                       form=('Nblts', 'Nspws', 'Nfreqs', 'Npols'),
-                                       expected_type=np.complex)
+        self._data_array = uvp.UVParameter('data_array', description=desc,
+                                           form=('Nblts', 'Nspws', 'Nfreqs', 'Npols'),
+                                           expected_type=np.complex)
 
-        self._vis_units = UVParameter('vis_units',
-                                      description='Visibility units, options '
-                                                  '["uncalib","Jy","K str"]',
-                                      form='str')
+        self._vis_units = uvp.UVParameter('vis_units',
+                                          description='Visibility units, options '
+                                                      '["uncalib","Jy","K str"]',
+                                          form='str')
 
         desc = ('number of data points averaged into each data element, '
                 'type = int, same shape as data_array')
-        self._nsample_array = UVParameter('nsample_array', description=desc,
-                                          form=('Nblts', 'Nspws', 'Nfreqs', 'Npols'),
-                                          expected_type=(np.float, np.int))
+        self._nsample_array = uvp.UVParameter('nsample_array', description=desc,
+                                              form=('Nblts', 'Nspws', 'Nfreqs', 'Npols'),
+                                              expected_type=(np.float, np.int))
 
         desc = 'boolean flag, True is flagged, same shape as data_array.'
-        self._flag_array = UVParameter('flag_array', description=desc,
-                                       form=('Nblts', 'Nspws', 'Nfreqs', 'Npols'),
-                                       expected_type=np.bool)
+        self._flag_array = uvp.UVParameter('flag_array', description=desc,
+                                           form=('Nblts', 'Nspws', 'Nfreqs', 'Npols'),
+                                           expected_type=np.bool)
 
-        self._Nspws = UVParameter('Nspws', description='number of spectral windows '
-                                  '(ie non-contiguous spectral chunks)')
+        self._Nspws = uvp.UVParameter('Nspws', description='number of spectral windows '
+                                      '(ie non-contiguous spectral chunks)')
 
-        self._spw_array = UVParameter('spw_array', description='array of spectral window '
-                                      'numbers', form=('Nspws',))
+        self._spw_array = uvp.UVParameter('spw_array', description='array of spectral window '
+                                          'numbers', form=('Nspws',))
 
         desc = ('Projected baseline vectors relative to phase center, ' +
                 '(3,Nblts), units meters')
-        self._uvw_array = UVParameter('uvw_array', description=desc,
-                                      form=(3, 'Nblts'), expected_type=np.float,
-                                      sane_vals=(1e-3, 1e8), tols=.001)
+        self._uvw_array = uvp.UVParameter('uvw_array', description=desc,
+                                          form=(3, 'Nblts'), expected_type=np.float,
+                                          sane_vals=(1e-3, 1e8), tols=.001)
 
         desc = ('array of times, center of integration, dimension (Nblts), ' +
                 'units Julian Date')
-        self._time_array = UVParameter('time_array', description=desc,
-                                       form=('Nblts',), expected_type=np.float,
-                                       tols=1e-3 / (60.0 * 60.0 * 24.0))  # 1 ms in days
+        self._time_array = uvp.UVParameter('time_array', description=desc,
+                                           form=('Nblts',), expected_type=np.float,
+                                           tols=1e-3 / (60.0 * 60.0 * 24.0))  # 1 ms in days
 
         desc = ('array of lsts, center of integration, dimension (Nblts), ' +
                 'units radians')
-        self._lst_array = UVParameter('lst_array', description=desc, form=('Nblts',),
-                                      expected_type=np.float,
-                                      tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 ms in radians
+        self._lst_array = uvp.UVParameter('lst_array', description=desc, form=('Nblts',),
+                                          expected_type=np.float,
+                                          tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 ms in radians
 
         desc = ('array of first antenna indices, dimensions (Nblts), '
                 'type = int, 0 indexed')
-        self._ant_1_array = UVParameter('ant_1_array', description=desc, form=('Nblts',))
+        self._ant_1_array = uvp.UVParameter('ant_1_array', description=desc, form=('Nblts',))
         desc = ('array of second antenna indices, dimensions (Nblts), '
                 'type = int, 0 indexed')
-        self._ant_2_array = UVParameter('ant_2_array', description=desc, form=('Nblts',))
+        self._ant_2_array = uvp.UVParameter('ant_2_array', description=desc, form=('Nblts',))
 
         desc = ('array of baseline indices, dimensions (Nblts), '
                 'type = int; baseline = 2048 * (ant2+1) + (ant1+1) + 2^16 '
                 '(may this break casa?)')
-        self._baseline_array = UVParameter('baseline_array', description=desc, form=('Nblts',))
+        self._baseline_array = uvp.UVParameter('baseline_array', description=desc, form=('Nblts',))
 
         # this dimensionality of freq_array does not allow for different spws
         # to have different dimensions
         desc = 'array of frequencies, dimensions (Nspws,Nfreqs), units Hz'
-        self._freq_array = UVParameter('freq_array', description=desc,
-                                       form=('Nspws', 'Nfreqs'),
-                                       expected_type=np.float,
-                                       tols=1e-3)  # mHz
+        self._freq_array = uvp.UVParameter('freq_array', description=desc,
+                                           form=('Nspws', 'Nfreqs'),
+                                           expected_type=np.float,
+                                           tols=1e-3)  # mHz
 
         desc = ('array of polarization integers (Npols). '
                 'AIPS Memo 117 says: stokes 1:4 (I,Q,U,V);  '
                 'circular -1:-4 (RR,LL,RL,LR); linear -5:-8 (XX,YY,XY,YX)')
-        self._polarization_array = UVParameter('polarization_array',
-                                               description=desc, form=('Npols',))
+        self._polarization_array = uvp.UVParameter('polarization_array',
+                                                   description=desc, form=('Npols',))
 
-        self._integration_time = UVParameter('integration_time',
-                                             description='length of the integration (s)',
-                                             expected_type=np.float, tols=1e-3)  # 1 ms
-        self._channel_width = UVParameter('channel_width',
-                                          description='width of channel (Hz)',
-                                          expected_type=np.float,
-                                          tols=1e-3)  # 1 mHz
+        self._integration_time = uvp.UVParameter('integration_time',
+                                                 description='length of the integration (s)',
+                                                 expected_type=np.float, tols=1e-3)  # 1 ms
+        self._channel_width = uvp.UVParameter('channel_width',
+                                              description='width of channel (Hz)',
+                                              expected_type=np.float,
+                                              tols=1e-3)  # 1 mHz
 
         # --- observation information ---
-        self._object_name = UVParameter('object_name', description='source or field '
-                                        'observed (string)', form='str')
-        self._telescope_name = UVParameter('telescope_name', description='name of telescope '
-                                           '(string)', form='str')
-        self._instrument = UVParameter('instrument', description='receiver or backend.',
-                                       form='str')
-        self._latitude = AngleParameter('latitude',
-                                        description='latitude of telescope, units radians',
-                                        expected_type=np.float,
-                                        tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
-        self._longitude = AngleParameter('longitude',
-                                         description='longitude of telescope, units radians',
+        self._object_name = uvp.UVParameter('object_name', description='source or field '
+                                            'observed (string)', form='str')
+        self._telescope_name = uvp.UVParameter('telescope_name',
+                                               description='name of telescope '
+                                               '(string)', form='str')
+        self._instrument = uvp.UVParameter('instrument', description='receiver or backend.',
+                                           form='str')
+        self._latitude = uvp.AngleParameter('latitude',
+                                            description='latitude of telescope, units radians',
+                                            expected_type=np.float,
+                                            tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
+        self._longitude = uvp.AngleParameter('longitude',
+                                             description='longitude of telescope, units radians',
+                                             expected_type=np.float,
+                                             tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
+        self._altitude = uvp.UVParameter('altitude',
+                                         description='altitude of telescope, units meters',
                                          expected_type=np.float,
-                                         tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
-        self._altitude = UVParameter('altitude',
-                                     description='altitude of telescope, units meters',
-                                     expected_type=np.float,
-                                     tols=1e-3)  # 1 mm
-        self._history = UVParameter('history', description='string of history, units English',
-                                    form='str')
+                                         tols=1e-3)  # 1 mm
+        self._history = uvp.UVParameter('history', description='string of history, units English',
+                                        form='str')
 
         desc = ('epoch year of the phase applied to the data (eg 2000.)')
-        self._phase_center_epoch = UVParameter('phase_center_epoch', description=desc,
-                                               expected_type=np.float)
+        self._phase_center_epoch = uvp.UVParameter('phase_center_epoch', description=desc,
+                                                   expected_type=np.float)
 
         self._is_phased = UVParameter('is_phased', required=True,
                                         expected_type=bool,
@@ -276,119 +142,119 @@ class UVData(object):
         # --- antenna information ----
         desc = ('number of antennas with data present. May be smaller ' +
                 'than the number of antennas in the array')
-        self._Nants_data = UVParameter('Nants_data', description=desc)
+        self._Nants_data = uvp.UVParameter('Nants_data', description=desc)
         desc = ('number of antennas in the array. May be larger ' +
                 'than the number of antennas with data')
-        self._Nants_telescope = UVParameter('Nants_telescope', description=desc)
+        self._Nants_telescope = uvp.UVParameter('Nants_telescope', description=desc)
         desc = ('list of antenna names, dimensions (Nants_telescope), '
                 'indexed by self.ant_1_array, self.ant_2_array, '
                 'self.antenna_indices. There must be one '
                 'entry here for each unique entry in self.ant_1_array and '
                 'self.ant_2_array, but there may be extras as well.')
-        self._antenna_names = UVParameter('antenna_names', description=desc,
-                                          form=('Nants_telescope',),
-                                          expected_type=str)
+        self._antenna_names = uvp.UVParameter('antenna_names', description=desc,
+                                              form=('Nants_telescope',),
+                                              expected_type=str)
 
         desc = ('integer index into antenna_names, dimensions '
                 '(Nants_telescope). There must be one '
                 'entry here for each unique entry in self.ant_1_array and '
                 'self.ant_2_array, but there may be extras as well.')
-        self._antenna_indices = UVParameter('antenna_indices', description=desc,
-                                            form=('Nants_telescope',))
+        self._antenna_indices = uvp.UVParameter('antenna_indices', description=desc,
+                                                form=('Nants_telescope',))
 
         # -------- extra, non-required parameters ----------
         desc = ('any user supplied extra keywords, type=dict')
-        self._extra_keywords = ExtraKeywordParameter('extra_keywords',
-                                                     description=desc)
+        self._extra_keywords = uvp.ExtraKeywordParameter('extra_keywords',
+                                                         description=desc)
 
-        self._dateobs = UVParameter('dateobs', required=False,
-                                    description='date of observation')
+        self._dateobs = uvp.UVParameter('dateobs', required=False,
+                                        description='date of observation')
 
         desc = ('coordinate frame for antenna positions '
                 '(eg "ITRF" -also google ECEF). NB: ECEF has x running '
                 'through long=0 and z through the north pole')
-        self._xyz_telescope_frame = UVParameter('xyz_telescope_frame',
-                                                required=False, description=desc,
-                                                spoof_val='ITRF', form='str')
+        self._xyz_telescope_frame = uvp.UVParameter('xyz_telescope_frame',
+                                                    required=False, description=desc,
+                                                    spoof_val='ITRF', form='str')
 
-        self._x_telescope = UVParameter('x_telescope', required=False,
-                                        description='x coordinates of array '
-                                                    'center in meters in coordinate frame',
-                                        spoof_val=0,
-                                        tols=1e-3)  # 1 mm
-        self._y_telescope = UVParameter('y_telescope', required=False,
-                                        description='y coordinates of array '
-                                                    'center in meters in coordinate frame',
-                                        spoof_val=0,
-                                        tols=1e-3)  # 1 mm
-        self._z_telescope = UVParameter('z_telescope', required=False,
-                                        description='z coordinates of array '
-                                                    'center in meters in coordinate frame',
-                                        spoof_val=0,
-                                        tols=1e-3)  # 1 mm
+        self._x_telescope = uvp.UVParameter('x_telescope', required=False,
+                                            description='x coordinates of array '
+                                                        'center in meters in coordinate frame',
+                                            spoof_val=0,
+                                            tols=1e-3)  # 1 mm
+        self._y_telescope = uvp.UVParameter('y_telescope', required=False,
+                                            description='y coordinates of array '
+                                                        'center in meters in coordinate frame',
+                                            spoof_val=0,
+                                            tols=1e-3)  # 1 mm
+        self._z_telescope = uvp.UVParameter('z_telescope', required=False,
+                                            description='z coordinates of array '
+                                                        'center in meters in coordinate frame',
+                                            spoof_val=0,
+                                            tols=1e-3)  # 1 mm
         desc = ('array giving coordinates of antennas relative to '
                 '{x,y,z}_telescope in the same frame, (Nants_telescope, 3)')
-        self._antenna_positions = AntPositionParameter('antenna_positions',
-                                                       required=False,
-                                                       description=desc,
-                                                       form=('Nants_telescope', 3),
-                                                       tols=1e-3)  # 1 mm
+        self._antenna_positions = uvp.AntPositionParameter('antenna_positions',
+                                                           required=False,
+                                                           description=desc,
+                                                           form=('Nants_telescope', 3),
+                                                           tols=1e-3)  # 1 mm
 
         desc = ('ra of zenith. units: radians, shape (Nblts)')
-        self._zenith_ra = AngleParameter('zenith_ra', required=False,
-                                         description=desc,
-                                         form=('Nblts',),
-                                         tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
+        self._zenith_ra = uvp.AngleParameter('zenith_ra', required=False,
+                                             description=desc,
+                                             form=('Nblts',),
+                                             tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
 
         desc = ('dec of zenith. units: radians, shape (Nblts)')
         # in practice, dec of zenith will never change; does not need to be shape Nblts
-        self._zenith_dec = AngleParameter('zenith_dec', required=False,
-                                          description=desc,
-                                          form=('Nblts',),
-                                          tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
+        self._zenith_dec = uvp.AngleParameter('zenith_dec', required=False,
+                                              description=desc,
+                                              form=('Nblts',),
+                                              tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
 
         desc = ('right ascension of phase center (see uvw_array), '
                 'units radians')
-        self._phase_center_ra = AngleParameter('phase_center_ra', required=False,
-                                               description=desc,
-                                               tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
+        self._phase_center_ra = uvp.AngleParameter('phase_center_ra', required=False,
+                                                   description=desc,
+                                                   tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
 
         desc = ('declination of phase center (see uvw_array), '
                 'units radians')
-        self._phase_center_dec = AngleParameter('phase_center_dec', required=False,
-                                                description=desc,
-                                                tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
+        self._phase_center_dec = uvp.AngleParameter('phase_center_dec', required=False,
+                                                    description=desc,
+                                                    tols=2 * np.pi * 1e-3 / (60.0 * 60.0 * 24.0))  # 1 mas in radians
 
         # --- other stuff ---
         # the below are copied from AIPS memo 117, but could be revised to
         # merge with other sources of data.
-        self._GST0 = UVParameter('GST0', required=False,
-                                 description='Greenwich sidereal time at '
-                                             'midnight on reference date',
-                                 spoof_val=0.0)
-        self._RDate = UVParameter('RDate', required=False,
-                                  description='date for which the GST0 or '
-                                              'whatever... applies',
-                                  spoof_val='')
-        self._earth_omega = UVParameter('earth_omega', required=False,
-                                        description='earth\'s rotation rate '
-                                                    'in degrees per day',
-                                        spoof_val=360.985)
-        self._DUT1 = UVParameter('DUT1', required=False,
-                                 description='DUT1 (google it) AIPS 117 '
-                                             'calls it UT1UTC',
-                                 spoof_val=0.0)
-        self._TIMESYS = UVParameter('TIMESYS', required=False,
-                                    description='We only support UTC',
-                                    spoof_val='UTC', form='str')
+        self._GST0 = uvp.UVParameter('GST0', required=False,
+                                     description='Greenwich sidereal time at '
+                                                 'midnight on reference date',
+                                     spoof_val=0.0)
+        self._RDate = uvp.UVParameter('RDate', required=False,
+                                      description='date for which the GST0 or '
+                                                  'whatever... applies',
+                                      spoof_val='')
+        self._earth_omega = uvp.UVParameter('earth_omega', required=False,
+                                            description='earth\'s rotation rate '
+                                                        'in degrees per day',
+                                            spoof_val=360.985)
+        self._DUT1 = uvp.UVParameter('DUT1', required=False,
+                                     description='DUT1 (google it) AIPS 117 '
+                                                 'calls it UT1UTC',
+                                     spoof_val=0.0)
+        self._TIMESYS = uvp.UVParameter('TIMESYS', required=False,
+                                        description='We only support UTC',
+                                        spoof_val='UTC', form='str')
 
         desc = ('FHD thing we do not understand, something about the time '
                 'at which the phase center is normal to the chosen UV plane '
                 'for phasing')
-        self._uvplane_reference_time = UVParameter('uvplane_reference_time',
-                                                   required=False,
-                                                   description=desc,
-                                                   spoof_val=0)
+        self._uvplane_reference_time = uvp.UVParameter('uvplane_reference_time',
+                                                       required=False,
+                                                       description=desc,
+                                                       spoof_val=0)
 
         for p in self.parameter_iter():
             this_param = getattr(self, p)
@@ -414,7 +280,7 @@ class UVData(object):
         param_list = []
         for a in attribute_list:
             attr = getattr(self, a)
-            if isinstance(attr, UVParameter):
+            if isinstance(attr, uvp.UVParameter):
                 param_list.append(a)
         for a in param_list:
             yield a
@@ -425,7 +291,7 @@ class UVData(object):
         required_list = []
         for a in attribute_list:
             attr = getattr(self, a)
-            if isinstance(attr, UVParameter):
+            if isinstance(attr, uvp.UVParameter):
                 if attr.required:
                     required_list.append(a)
         for a in required_list:
@@ -437,7 +303,7 @@ class UVData(object):
         extra_list = []
         for a in attribute_list:
             attr = getattr(self, a)
-            if isinstance(attr, UVParameter):
+            if isinstance(attr, uvp.UVParameter):
                 if not attr.required:
                     extra_list.append(a)
         for a in extra_list:
